imx6: fix GW52xx usbotg select pinmux
GPIO2 is used to select the routing of the USBOTG bus between the
Front Panel connector and the PCIe Socket.  By default we want this
to route to the front panel which this patch achieves.

To route USBOTG to the PCIe socket from Linux, export the GPIO and drive it
high:
  echo 2 > /sys/class/gpio/export
  echo out > /sys/class/gpio/gpio2/direction
  echo 1 > /sys/class/gpio/gpio2/value  ;# high to route to PCIe socket J8
  echo 0 > /sys/class/gpio/gpio2/value  ;# low to route to front panel OTG

Note that when routing USB to the PCIe socket J8 you must have a USBOTG
micro-A cable connected which grounds the ID pin to put the OTG controller
into host mode.  This will not be necessary on revision B+ of the GW52xx as
the grounding of this signal will be done on-board when GPIO2 is driven high
to route USBOTG to J8.