ventana: do not de-assert PCIE_RST#
PCIE_RST# should not be de-asserted until after the clock is valid.  Ventana
boards use the following PCI clock scenarios:
 - GW51XX - clock provided by CPU LVDS CLK1 out
 - GW52XX - clocks provided by CPU LVDS CLK1 out via SI53154 clock buffer
 - GW53XX - clocks provided by CPU LVDS CLK1 out via SI53156 clock buffer
 - GW54XX - clocks provided by SI52147 clock generator

In all cases except the GW54XX with a clock generator, the PCI clocks are not
valid until the PCIe host controller driver loads in which case the driver
enables the clock and asserts PCIE_RST# appropriately and thus PCIE_RST# should
not be de-asserted by the bootloader at all (until/unless the bootloader
supports PCIe host controller).  We'll stay away from doing it on the GW54XX
as well for consistancy even though its clocks are valid by the time the
bootloader is running.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>