ventana: add DIO pad config to hwconfig
The hwconfig env variable can be used to configure the pinmux and padctrl
registers selecting the internal routing of the pad (gpio vs pwm where
applicable) and the IO cell configuration (pullup/down configuration, drive
strength, etc).

This patch adds the ability to configure the padctrl with the 'padctrl'
subarg.  The default configuration if not specified is a 100k pullup,
medium speed, 40ohm drive strength, hysterisis, and fast slew rate.

The padctrl subarg must be a hex string (leading 0x) that conforms to the
pad_ctrl register defintion described in the IMX6 reference manual.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>