usb storage: set env variable to host controller no
In order to support boot scripts that dynamically create kernel arguments
that describe the boot device, a method is needed to determine what USB
host controller number is associated with current usb storage device set
via 'usb dev <n>'.

This patch dynamically sets the 'usbdevport' env variable according to the
instance number of the EHCI host controller when the usb storage device is
set with the 'usb dev <n>' command.

This is currently being used to be able to pass a kernel command-line argument
the the IMX Android kernel to tell it which USB host device to boot from:
fsl-ehci.0 (OTG) or fsl-ehci.1 (EHCI).

Signed-off-by: Tim Harvey <tharvey@gateworks.com>