imx: ventana: switch to SPL
Switch to an SPL image. The SPL for Ventana does the following:
 - setup i2c and read the factory programmed EEPROM to obtain DRAM config
   and model for board-specific calibration data
 - configure DRAM per CPU/size/layout/devices/calibration
 - load u-boot.img from NAND and jump to it

This allows for a single SPL+u-boot.img to replace the previous multiple boa
configurations.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>

v3:
- re-ordered calls in board_init_f

v2:
- remove missing/unnecessary include
- revert mtdparts change
- use imx_ddr_size() which uses MMDC configuration to determine mem size
  explicitly
- add support for MX6SOLO and MX6DUAL
- set CS0_END for 4GB so get_ram_size() works
- updated DDR3 calibration values for ventana boards
- fixed build issue - only compile spl if doing spl build
- fixed line length issue in README
- remove CONFIG_SPL* conditions and conditionally compile instead
- removed prints for CPU type and DRAM size/width - uboot will print these
  later
- removed unused gw_ventana_spl.cfg
- use common read_eeprom function
- added MMC support to SPL

imx: ventana: fix SPL image load

arch_cpu_init() calls mxs_dma_init() to start APBH DMA and for reasons unknown
if this is not done between initializing the MMDC and clearing the BSS
loading NAND SPL will fail.

I've added debugging down to loading NAND and the image gets loaded properly
in memory, but the entry point is somehow resolved as 0. The spl_image struct
is a global so it makes sense its related to the BSS - perhaps there needs
to be some cache coherence wmb there.

merge with SPL