ventana: add initial environment adjustment for setting pcidisable
The GW5520 is currently the only Ventana board that has use of PCI in the
bootloader. We will default PCI to be disabled for all other boards for
the time being. This is mainly because we are still supporing some kernels
which have an issue in their pci driver initializing which will cause a hang
if PCI was enabled in the bootloader.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>