Revert "ventana: add initial environment adjustment for setting pcidisable"
This reverts commit a27b08cf20823c2ae3f3e7d8f303cd4abf2845a3.

modern Linux kernels now have workarounds for handling PCI being enabled
by the bootloader and this is no longer needed.

Removing this and enabling PCI in U-Boot also resolves an issue with the
GW16103 and GW16112 miniPCIe expansion mezzanines as they are held in reset
until PCI enumeration making them unusable in the bootloader.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>