imx: ventana: add dt fixups for watchdog external reset
Added removal of the ext-reset-output property in the wdog node for board
revisions that pre-date the addition of the external watchdog reset signal.

This property is a recent addition to mainline linux kernel in order to
specify that the IMX watchdog external reset should be used instead of the
internal chip-level reset.

Additionally convenience functions for disabling a dt node and deleting a
property are of notes that potentially may not exist are added and used.

TODO: add board-level support for specifying wdog1 vs wdog2 and internal vs
external reset in imx_watchdog driver.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>