imx: ventana: bump 1GHz ARM voltage rail for GW53xx
The GW5300 revision A-D has insufficient trace width routing for VDD_ARM_IN
from the PMIC to the IMX SoC which can cause instability on IMX6Q when running
in ldo-bypass mode and switching to the 1GHz CPU frequency.

This patch bumps the 1GHz ARM operating point in the device-tree for such
kernels to resolve this.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>