imx: ventana: fix PMIC rail adjustments for IMX6SDL in ldo-bypass mode
The IMX6SDL has a VDD_SOC max of 1.21v when operating in ldo-bypass mode.
While the kernel technically should be in charge of setting the rails and
bypassing the internal LDO regulators if implementing ldo-bypass mode, it
is impossible to meet the VDD_SOC_IN min and max requirements while switching.

It has been found that exceeding the max of 1.21v in ldo-bypass mode
causes PCIe aborts hanging the system and not meeting the min of 1.275v
in ldo-enabled mode can produce other issues.

As a work-around we will examine the device-tree bindings during ft_board_setup
to determine if we are booting a configured for ldo-bypass and will
adjust the PMIC rails, and bypass the LDO's before booting the kernel.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>