ventana: use continuous PWM mode and 1MHz for LTC3676 PMIC regulators
It has been found that the default pulse-skipping mode of the LTC3676
regulators produces noise that the IMX6 is especially susceptible to when
run in ldo-bypass mode.

Switch to continuous PWM mode to provide a cleaner rail, as well as 1MHz
switching frequency for EMI reduction. There is no noticeable drop in
efficiency between pulse-skipping mode and PWM mode in the Ventana IMX6
use case.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>