imx: ventana: update imx wdog external reset dt property
The devicetree property that was accepted upstream in the imx wdog driver
was 'fsl,ext-reset-output' not 'ext-reset-output'.

Therefore we will remove both of these properties to account for the mainline
kernels as well as the downstream vendor kernels.

This resolves issues with soft reboot failing to reset the IMX6 SoC on older
revision boards using Linux v4.9+

Signed-off-by: Tim Harvey <tharvey@gateworks.com>