imx: ventana: gw5909: remove disable of rs232
Ventana boards always route UART2 (serial console) to the JTAG connector
as well as to an RS232 transceiver which we always leave enabled by default.

Note that this causes a conflict between JTAG and RS232 and the JTAG
devices have a stronger driver, therefore if you wish to use RS232 physically
disconnect the JTAG serial connection.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>