imx: ventana: remove FEC phy-reset gpio from device-tree
The Linux FEC driver will reset the PHY if the phy-reset-gpios prop
defines a reset pin. This causes all the PHY configuration we perform
in the bootloader to be undone such as LED configuration, RGMII internal
delay, and clock configuration. The Linux PHY drivers typically don't
handle LED configuration so its best to just keep them from resetting
the PHY.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>