imx: fix boot device detection following soft reset (ie bmode)
To determine the boot device after power-on-reset you must look at
SRC_SBMR1 but if you want to determine this after a soft-reset
you should look at SRC_GPR9 which is written to by the boot rom.
To make this decision SRC_GPR10[28] is examined which is also
set by the boot rom on soft reset.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>